+++ lib/system.dart
@@ -3,5 +3,7 @@ import 'package:tecs/world.dart';
 abstract class System<T> {
   late final World world;
   late final String tag;
+
+  void init() {}
   void update(T args);
 }