+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.4
+
+* Querying will return the list of rows directly instead of capsulating them in a class.
+* Ability to query with parameters has been added. ``queryWithParams``, ``queryCountWithParams`` 
+* Helper functions added for creating custom queries. ``componentID``, ``componentColumn``, ``archetypesWith``
+
 ## 1.0.3
 
 * Ability to create entities with multiple components, in bulk.